pyproject.toml
[project]
name = "pychub"
version = "0.0.1"
description = "Python CLI for Context Hub"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Context Hub Contributors" }]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "PyYAML>=6.0",
    "httpx>=0.27",
    "pydantic>=2.7",
    "posthog>=3.5",
]

[project.scripts]
pychub = "pychub.cli:main"
pychub-mcp = "pychub.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pychub"]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["pytests"]
pythonpath = ["src"]
